+++ app/src/app_base.h
@@ -299,6 +299,10 @@ struct st_app_rx_audio_session {
   int stat_frame_total_received;
   uint64_t stat_frame_first_rx_time;
   double expect_fps;
+
+  bool enable_timing_parser_meta;
+  uint32_t stat_compliant_result[ST_RX_TP_COMPLIANT_MAX];
+  int32_t ipt_max;
 };
 
 struct st_app_rx_anc_session {
@@ -508,6 +512,7 @@ struct st_app_context {
   bool tx_copy_once;
   bool app_thread;
   bool enable_timing_parser;
+  bool enable_timing_parser_meta;
   bool tx_display;
   bool rx_display;
   uint16_t rx_burst_size;